Add useStorageRevision and migrate remaining inline storage listeners
SpeedKatakanaHeatmap, SpeedKatakanaBreakdown (revision/remount pattern)
and ItemCountBadge's bookmark counter now use the shared hooks instead
of hand-rolled storage-event effects.
